feat: add fernspielapparat discovery
Add functionality to start up a fernspielapparat server on a random port. If a runtime is found on the path, it is preferred. If none is found, checks if one was previously downloaded and uses that. If nothing found, asks GitHub for the latest release and then downloads and extracts it to a user directory (Application Support on Mac).