chore: add electron-forge to build release packages
Add `electron-forge` to the project by adding the following runtime dependencies to `package.json`:
* `electron-compile`,
* `targz` (used to be dev dependency).
Also add the following compile time dependencies:
* `@electron-forge/maker-zip`,
* `electron`,
* `electron-forge`,
* `electron-prebuilt-compile`.
The configuration for the electron forge build is contained in `forge.config.js`
and is set up to build `.zip` distributables for all platforms.
On Mac, it additionally builds a `.dmg` distributable.