chore: add webpack support
`yarn start` now spins up a webpack server, making ES6 features available for both the main and the server process.
Live reload is available for the renderer process only.
`preload.js` did not really work together with webpack, so activated node integration for the renderer process and moved the logic to `renderer.js`.