fix: download correct arch on x64
Fix a bug on x86 linux systems where the ARM version would be downloaded since both contain `-linux-`. Fix by always downloading for x86 through an adaptation of the regex. This has the side effect of breaking this for RPi, but we do not intend to release for this platform as of now.